# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 11 Swagger UI API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Orders",
                "description": "Mendapatkan semua orders",
                "operationId": "orders",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store Order",
                "description": "Menambahkan pesanan baru",
                "operationId": "storeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_customer": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "pending"
                                    },
                                    "order_items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_product": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": "1000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show Order",
                "description": "Mendapatkan detail pelanggan berdasarkan ID",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete Order",
                "description": "Menghapus data order",
                "operationId": "deleteOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}